restructure token relations graph model and components
