+++ packages/frontend/src/components/projects/sections/interop/onchain-deployments/columns.tsx
@@ -10,7 +10,7 @@ import { ProjectIconList } from '~/components/ProjectIconList'
 import type { BasicTableRow } from '~/components/table/BasicTable'
 import { IndexCell } from '~/components/table/cells/IndexCell'
 import { InteropNoDataBadge } from '~/pages/interop/components/InteropNoDataBadge'
-import type { InteropTokenOnchainDeploymentsRow } from './InteropTokenOnchainDeploymentsSection'
+import type { InteropTokenRelationsDeployment } from '~/server/features/layer2s/interop/token/getInteropTokenRelationsGraph'
 
 const UNSUPPORTED_CHAIN_TOOLTIP =
   "The information is not available as this deployment is on a chain we don't fully support."
@@ -19,7 +19,7 @@ const NO_TRANSFER_TIME_TOOLTIP =
 const NO_MINTERS_TOOLTIP =
   'No known bridge has been observed minting this deployment. It is likely the locked or natively issued side.'
 
-export type DeploymentRow = InteropTokenOnchainDeploymentsRow & BasicTableRow
+export type DeploymentRow = InteropTokenRelationsDeployment & BasicTableRow
 const columnHelper = createColumnHelper<DeploymentRow>()
 export const interopTokenOnchainDeploymentsColumns = [
   columnHelper.accessor((_, index) => index + 1, {
