+++ packages/frontend/src/server/features/layer2s/interop/token/getInteropTokenEntry.test.ts
@@ -1,5 +1,4 @@
-import type { InteropPlugin, Project } from '@l2beat/config'
-import { Address32, assert, ProjectId, UnixTime } from '@l2beat/shared-pure'
+import { Address32, assert } from '@l2beat/shared-pure'
 import { expect } from 'earl'
 import { getInteropTokenEntry } from './getInteropTokenEntry'
 import type { InteropTokenOnchainDeployment } from './getInteropTokenOnchainDeployments'
@@ -11,68 +10,6 @@ const NO_RELATIONS: InteropTokenRelations = {
 }
 
 describe(getInteropTokenEntry.name, () => {
-  it('maps minting plugins to deduplicated project summaries', () => {
-    const entry = getInteropTokenEntry(
-      'circle-usdc',
-      [],
-      [],
-      [
-        project({
-          id: 'zeta',
-          interopName: 'Zeta bridge',
-          plugins: [{ plugin: 'cctp-v2', bridgeType: 'burnAndMint' }],
-        }),
-        project({
-          id: 'alpha',
-          interopName: 'Alpha bridge',
-          plugins: [
-            { plugin: 'ccip', bridgeType: 'burnAndMint' },
-            { plugin: 'cctp-v2', bridgeType: 'burnAndMint' },
-          ],
-        }),
-      ],
-      [
-        deployment({
-          mintingPlugins: [
-            {
-              plugin: 'ccip',
-              bridgeType: 'burnAndMint',
-              relatedChain: 'ethereum',
-            },
-            {
-              plugin: 'cctp-v2',
-              bridgeType: 'burnAndMint',
-              relatedChain: 'ethereum',
-            },
-          ],
-        }),
-      ],
-      NO_RELATIONS,
-    )
-
-    const section = entry.sections.find(
-      (section) => section.type === 'InteropTokenOnchainDeploymentsSection',
-    )
-    assert(section?.type === 'InteropTokenOnchainDeploymentsSection')
-
-    const minters = section.props.deployments[0]?.minters
-    assert(minters)
-    expect(minters).toEqual([
-      {
-        id: ProjectId('alpha'),
-        name: 'Alpha bridge',
-        iconUrl: '/icons/alpha.png',
-        href: '/interop/protocols/alpha',
-      },
-      {
-        id: ProjectId('zeta'),
-        name: 'Zeta bridge',
-        iconUrl: '/icons/zeta.png',
-        href: '/interop/protocols/zeta',
-      },
-    ])
-  })
-
   it('takes deployment stats from the pair stats and sorts by volume', () => {
     const ethereum = deployment({ chain: 'ethereum', address: '0xe1' })
     const base = deployment({ chain: 'base', address: '0xb1' })
@@ -163,54 +100,6 @@ describe(getInteropTokenEntry.name, () => {
       })?.nodes.map((node) => node.deployments.length),
     ).toEqual([2])
   })
-
-  it('skips manually added relations when resolving minters', () => {
-    const entry = getInteropTokenEntry(
-      'ether',
-      [],
-      [],
-      [
-        project({
-          id: 'zeta',
-          interopName: 'Zeta bridge',
-          plugins: [{ plugin: 'cctp-v2', bridgeType: 'burnAndMint' }],
-        }),
-      ],
-      [
-        deployment({
-          mintingPlugins: [
-            {
-              plugin: 'manual',
-              bridgeType: 'lockAndMint',
-              relatedChain: 'ethereum',
-            },
-            {
-              plugin: 'cctp-v2',
-              bridgeType: 'burnAndMint',
-              relatedChain: 'ethereum',
-            },
-          ],
-        }),
-      ],
-      NO_RELATIONS,
-    )
-
-    const section = entry.sections.find(
-      (section) => section.type === 'InteropTokenOnchainDeploymentsSection',
-    )
-    assert(section?.type === 'InteropTokenOnchainDeploymentsSection')
-
-    const minters = section.props.deployments[0]?.minters
-    assert(minters)
-    expect(minters).toEqual([
-      {
-        id: ProjectId('zeta'),
-        name: 'Zeta bridge',
-        iconUrl: '/icons/zeta.png',
-        href: '/interop/protocols/zeta',
-      },
-    ])
-  })
 })
 
 function deployment(
@@ -223,28 +112,5 @@ function deployment(
     mintingPlugins: [],
     isSupported: true,
     ...override,
-  }
-}
-
-function project({
-  id,
-  interopName,
-  plugins,
-}: {
-  id: string
-  interopName: string
-  plugins: InteropPlugin[]
-}): Project<'interopConfig'> {
-  return {
-    id: ProjectId(id),
-    slug: id,
-    name: id,
-    shortName: undefined,
-    addedAt: UnixTime(0),
-    interopConfig: {
-      name: interopName,
-      plugins,
-      type: 'multichain',
-    },
   }
 }
